redrive failed link previews + renames
