+++ pillcity/daos/link_preview.py
@@ -2,12 +2,15 @@ from typing import Optional
 from mongoengine.errors import ValidationError
 from pillcity.models import LinkPreview, LinkPreviewState
 from pillcity.tasks.tasks import generate_link_preview
+from pillcity.utils.now import now_seconds
+
+LinkPreviewRefetchIntervalSeconds = 60
 
 
 def get_link_preview(url: str) -> Optional[LinkPreview]:
     try:
-        link_preview = LinkPreview.objects(url=url)
-        if not link_preview:
+        link_previews = LinkPreview.objects(url=url)
+        if not link_previews:
             new_link_preview = LinkPreview(
                 url=url,
                 state=LinkPreviewState.Fetching
@@ -15,6 +18,14 @@ def get_link_preview(url: str) -> Optional[LinkPreview]:
             new_link_preview.save()
             generate_link_preview.delay(url)
             return new_link_preview
-        return link_preview[0]
+        link_preview = link_previews[0]  # type: LinkPreview
+        now = now_seconds()
+        if link_preview.state == LinkPreviewState.Errored and \
+            (link_preview.last_refetched_seconds == 0 or
+             link_preview.last_refetched_seconds + LinkPreviewRefetchIntervalSeconds > now):
+            link_preview.last_refetched_seconds = now
+            link_preview.save()
+            generate_link_preview.delay(url)
+        return link_preview
     except (ValueError, ValidationError):
         return None
