+++ pillcity/tasks/tasks.py
@@ -42,7 +42,7 @@ def init_mongo():
         inited_mongo[0] = True
 
 
-@celery.task(rate_limit='8/m')
+@celery.task()
 def generate_link_preview(url: str):
     init_mongo()
     logger.info(f'Generating link preview for url {url}')
